requirements.txt
# Django
django==3.1.7

# Third party libs
pyjwt==2.0.1

# Linting
flake8==3.8.4